Use stdint.h in p00[4-6].c

+++ p004.c
@@ -1,7 +1,8 @@
+#include <stdint.h>
 #include <stdio.h>
 
-int reverse(int n) {
-	int result = 0;
+uint_fast32_t reverse(uint_fast32_t n) {
+	uint_fast32_t result = 0;
 	while(n > 0) {
 		result = result * 10 + n % 10;
 		n /= 10;
@@ -10,15 +11,15 @@ int reverse(int n) {
 }
 
 int main() {
-	int largest = 0;
-	for(int a = 999; a >= 100; a--) {
-		for(int b = 999; b >= a; b--) {
-			int product = a * b;
+	uint_fast32_t largest = 0;
+	for(uint_fast16_t a = 999; a >= 100; a--) {
+		for(uint_fast16_t b = 999; b >= a; b--) {
+			uint_fast32_t product = a * b;
 			if(product < largest)
 				break;
 			if(product == reverse(product))
 				largest = product;
 		}
 	}
-	printf("%d\n", largest);
+	printf("%lu\n", largest);
 }
