+++ p006.c
@@ -1,20 +1,21 @@
+#include <stdint.h>
 #include <stdio.h>
 
-unsigned int sum_of_squares(unsigned int n) {
-	unsigned int sum = 0;
-	for(unsigned int i = 1; i <= n; i++)
+uint_fast32_t sum_of_squares(uint_fast8_t n) {
+	uint_fast32_t sum = 0;
+	for(uint_fast16_t i = 1; i <= n; i++)
 		sum += i * i;
 
 	return sum;
 }
 
-unsigned int square_of_sum(unsigned int n) {
-	unsigned int sum = n * (n + 1) / 2;
+uint_fast32_t square_of_sum(uint_fast8_t n) {
+	uint_fast32_t sum = n * (n + 1) / 2;
 	return sum * sum;
 }
 
 int main() {
-	unsigned int n = 100;
-	unsigned int difference = square_of_sum(n) - sum_of_squares(n);
-	printf("%u\n", difference);
+	uint_fast8_t n = 100;
+	uint_fast32_t difference = square_of_sum(n) - sum_of_squares(n);
+	printf("%lu\n", difference);
 }
